server/requirement.txt
flask--2,2,2
flask_cors==4.0.0